Make device name selection case-insensitive (#4)
One of the things I did in the refactor was fiddle with device names,
sometimes changing the case. This PR makes the regex use case
insensitive matching.

A possible downside to this change is that it makes selection somewhat
less specific, but I don't think we want to live in a world where "Tiff"
vs "tiff" makes a meaningful difference and we should be the change we
want to see.