add an option for linking with /md (#14)
We link the Microsoft c runtime statically by default. For making a
micromanager add-on the dynamically linked runtime is preferred so we
need to be able to build core libs and runtime with /md.

This adds an option ACQUIRE_MSVC_USE_MD to the cmake build. Default is
to use /MT. Running

```
cmake -DACQUIRE_MSVC_USE_MD=ON ..
```

will configure to build with /MD.